Ghost kozak/add hangout function (#38)
* Added folder for Popup template to master rnd

Can't host site with this file. But can copy paste the code in it. Site is deployed to this link.
https://popup-temps.firebaseapp.com/

* Commented out inviting friends on the event page

Team has decided to go a different direction in terms of invitations.

* Merge conflict correction #1

* Added function, have not tested yet

This is used on the profile page of those invited to an event.

* Add Event function completed

Remeber, most things are sent to test folder. We will need to change code to event folder at some point